Ensure when syncing children we throw regardless of collection size
ImmutableList`1.RemoveRange does not enumerate its argument when the list is empty.
ImmutableSortedSet`1.Except *does* enumerate its argument even when the collection is empty.
This difference in behavior caused a test failure when preparing ProjectTree to change to use ordered Children instead of sorted children.

This change makes sure that every element in the argument is "realized" from the enumerable, and thus exceptions are reliably thrown.